+++ packages/google_maps_flutter/example/test_driver/google_map_inspector.dart
@@ -46,6 +46,10 @@ class GoogleMapInspector {
     return await _channel.invokeMethod<bool>('map#isScrollGesturesEnabled');
   }
 
+  Future<bool> isMyLocationEnabled() async {
+    return await _channel.invokeMethod<bool>('map#isMyLocationEnabled');
+  }
+
   Future<bool> isMyLocationButtonEnabled() async {
     return await _channel.invokeMethod<bool>('map#isMyLocationButtonEnabled');
   }