+++ packages/google_maps_flutter/example/test_driver/google_maps.dart
@@ -375,6 +375,34 @@ void main() {
     expect(secondVisibleRegion.contains(newCenter), isTrue);
   });
 
+  test('testMyLocation', () async {
+    final Key key = GlobalKey();
+    final Completer<GoogleMapInspector> inspectorCompleter =
+    Completer<GoogleMapInspector>();
+
+    await pumpWidget(Directionality(
+      textDirection: TextDirection.ltr,
+      child: GoogleMap(
+        key: key,
+        initialCameraPosition: _kInitialCameraPosition,
+        myLocationEnabled: true,
+        myLocationButtonEnabled: false,
+        onMapCreated: (GoogleMapController controller) {
+          final GoogleMapInspector inspector =
+          // ignore: invalid_use_of_visible_for_testing_member
+          GoogleMapInspector(controller.channel);
+          inspectorCompleter.complete(inspector);
+        },
+      ),
+    ));
+
+    final GoogleMapInspector inspector = await inspectorCompleter.future;
+    bool myLocationEnabled = await inspector.isMyLocationEnabled();
+    expect(myLocationEnabled, true);
+    bool myLocationButtonEnabled = await inspector.isMyLocationButtonEnabled();
+    expect(myLocationButtonEnabled, false);
+  });
+
   test('testMyLocationButtonToggle', () async {
     final Key key = GlobalKey();
     final Completer<GoogleMapInspector> inspectorCompleter =